amqp interop client
