+++ pkg/enqueue/Client/Amqp/AmqpDriver.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace  Enqueue\AmqpExt\Client;
+namespace  Enqueue\Client\Amqp;
 
-use Enqueue\AmqpExt\AmqpContext;
 use Enqueue\Client\Config;
 use Enqueue\Client\DriverInterface;
 use Enqueue\Client\Message;
 use Enqueue\Client\Meta\QueueMetaRegistry;
+use Interop\Amqp\AmqpContext;
 use Interop\Amqp\AmqpMessage;
 use Interop\Amqp\AmqpQueue;
 use Interop\Amqp\AmqpTopic;
