+++ pkg/enqueue/Tests/Client/Amqp/AmqpDriverTest.php
@@ -1,51 +1,44 @@
 <?php
 
-namespace Enqueue\AmqpExt\Tests\Client;
+namespace Enqueue\Tests\Client\Amqp;
 
-use Enqueue\AmqpExt\AmqpContext;
-use Enqueue\AmqpExt\Client\AmqpDriver;
-use Enqueue\AmqpExt\Client\RabbitMqDriver;
+use Enqueue\Client\Amqp\AmqpDriver;
 use Enqueue\Client\Config;
 use Enqueue\Client\DriverInterface;
 use Enqueue\Client\Message;
-use Enqueue\Client\MessagePriority;
 use Enqueue\Client\Meta\QueueMetaRegistry;
 use Enqueue\Test\ClassExtensionTrait;
-use Interop\Amqp\AmqpBind;
+use Interop\Amqp\AmqpContext;
+use Interop\Amqp\AmqpProducer;
+use Interop\Amqp\Impl\AmqpBind;
 use Interop\Amqp\Impl\AmqpMessage;
 use Interop\Amqp\Impl\AmqpQueue;
 use Interop\Amqp\Impl\AmqpTopic;
-use Interop\Queue\PsrProducer;
 use PHPUnit\Framework\TestCase;
 
-class RabbitMqDriverTest extends TestCase
+class AmqpDriverTest extends TestCase
 {
     use ClassExtensionTrait;
 
     public function testShouldImplementsDriverInterface()
     {
-        $this->assertClassImplements(DriverInterface::class, RabbitMqDriver::class);
-    }
-
-    public function testShouldExtendsAmqpDriverClass()
-    {
-        $this->assertClassExtends(AmqpDriver::class, RabbitMqDriver::class);
+        $this->assertClassImplements(DriverInterface::class, AmqpDriver::class);
     }
 
     public function testCouldBeConstructedWithRequiredArguments()
     {
-        new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            Config::create(),
+        new AmqpDriver(
+            $this->createAmqpContextMock(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
     }
 
     public function testShouldReturnConfigObject()
     {
-        $config = Config::create();
+        $config = $this->createDummyConfig();
 
-        $driver = new RabbitMqDriver($this->createPsrContextMock(), $config, $this->createDummyQueueMetaRegistry());
+        $driver = new AmqpDriver($this->createAmqpContextMock(), $config, $this->createDummyQueueMetaRegistry());
 
         $this->assertSame($config, $driver->getConfig());
     }
@@ -54,7 +47,7 @@ class RabbitMqDriverTest extends TestCase
     {
         $expectedQueue = new AmqpQueue('aName');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createQueue')
@@ -62,7 +55,7 @@ class RabbitMqDriverTest extends TestCase
             ->willReturn($expectedQueue)
         ;
 
-        $driver = new AmqpDriver($context, Config::create(), $this->createDummyQueueMetaRegistry());
+        $driver = new AmqpDriver($context, $this->createDummyConfig(), $this->createDummyQueueMetaRegistry());
 
         $queue = $driver->createQueue('aFooQueue');
 
@@ -76,7 +69,7 @@ class RabbitMqDriverTest extends TestCase
     {
         $expectedQueue = new AmqpQueue('aName');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createQueue')
@@ -84,7 +77,7 @@ class RabbitMqDriverTest extends TestCase
             ->willReturn($expectedQueue)
         ;
 
-        $driver = new AmqpDriver($context, Config::create(), $this->createDummyQueueMetaRegistry());
+        $driver = new AmqpDriver($context, $this->createDummyConfig(), $this->createDummyQueueMetaRegistry());
 
         $queue = $driver->createQueue('aBarQueue');
 
@@ -97,18 +90,16 @@ class RabbitMqDriverTest extends TestCase
         $transportMessage->setBody('body');
         $transportMessage->setHeaders(['hkey' => 'hval']);
         $transportMessage->setProperties(['key' => 'val']);
-        $transportMessage->setProperty('x-delay', '5678000');
         $transportMessage->setHeader('content_type', 'ContentType');
         $transportMessage->setHeader('expiration', '12345000');
-        $transportMessage->setHeader('priority', 3);
         $transportMessage->setMessageId('MessageId');
         $transportMessage->setTimestamp(1000);
         $transportMessage->setReplyTo('theReplyTo');
         $transportMessage->setCorrelationId('theCorrelationId');
 
-        $driver = new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            new Config('', '', '', '', '', '', ['delay_plugin_installed' => true]),
+        $driver = new AmqpDriver(
+            $this->createAmqpContextMock(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -120,7 +111,6 @@ class RabbitMqDriverTest extends TestCase
             'hkey' => 'hval',
             'content_type' => 'ContentType',
             'expiration' => '12345000',
-            'priority' => 3,
             'message_id' => 'MessageId',
             'timestamp' => 1000,
             'reply_to' => 'theReplyTo',
@@ -128,33 +118,13 @@ class RabbitMqDriverTest extends TestCase
         ], $clientMessage->getHeaders());
         $this->assertSame([
             'key' => 'val',
-            'x-delay' => '5678000',
         ], $clientMessage->getProperties());
         $this->assertSame('MessageId', $clientMessage->getMessageId());
         $this->assertSame(12345, $clientMessage->getExpire());
-        $this->assertSame(5678, $clientMessage->getDelay());
         $this->assertSame('ContentType', $clientMessage->getContentType());
         $this->assertSame(1000, $clientMessage->getTimestamp());
-        $this->assertSame(MessagePriority::HIGH, $clientMessage->getPriority());
         $this->assertSame('theReplyTo', $clientMessage->getReplyTo());
         $this->assertSame('theCorrelationId', $clientMessage->getCorrelationId());
-    }
-
-    public function testShouldThrowExceptionIfXDelayIsNotNumeric()
-    {
-        $transportMessage = new AmqpMessage();
-        $transportMessage->setProperty('x-delay', 'is-not-numeric');
-
-        $driver = new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            Config::create(),
-            $this->createDummyQueueMetaRegistry()
-        );
-
-        $this->expectException(\LogicException::class);
-        $this->expectExceptionMessage('x-delay header is not numeric. "is-not-numeric"');
-
-        $driver->createClientMessage($transportMessage);
     }
 
     public function testShouldThrowExceptionIfExpirationIsNotNumeric()
@@ -162,9 +132,9 @@ class RabbitMqDriverTest extends TestCase
         $transportMessage = new AmqpMessage();
         $transportMessage->setHeader('expiration', 'is-not-numeric');
 
-        $driver = new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            Config::create(),
+        $driver = new AmqpDriver(
+            $this->createAmqpContextMock(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -172,47 +142,6 @@ class RabbitMqDriverTest extends TestCase
         $this->expectExceptionMessage('expiration header is not numeric. "is-not-numeric"');
 
         $driver->createClientMessage($transportMessage);
-    }
-
-    public function testShouldThrowExceptionIfCantConvertTransportPriorityToClientPriority()
-    {
-        $transportMessage = new AmqpMessage();
-        $transportMessage->setHeader('priority', 'unknown');
-
-        $driver = new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            Config::create(),
-            $this->createDummyQueueMetaRegistry()
-        );
-
-        $this->expectException(\LogicException::class);
-        $this->expectExceptionMessage('Cant convert transport priority to client: "unknown"');
-
-        $driver->createClientMessage($transportMessage);
-    }
-
-    public function testShouldThrowExceptionIfCantConvertClientPriorityToTransportPriority()
-    {
-        $clientMessage = new Message();
-        $clientMessage->setPriority('unknown');
-
-        $context = $this->createPsrContextMock();
-        $context
-            ->expects($this->once())
-            ->method('createMessage')
-            ->willReturn(new AmqpMessage())
-        ;
-
-        $driver = new RabbitMqDriver(
-            $context,
-            Config::create(),
-            $this->createDummyQueueMetaRegistry()
-        );
-
-        $this->expectException(\LogicException::class);
-        $this->expectExceptionMessage('Given priority could not be converted to client\'s one. Got: unknown');
-
-        $driver->createTransportMessage($clientMessage);
     }
 
     public function testShouldConvertClientMessageToTransportMessage()
@@ -223,23 +152,21 @@ class RabbitMqDriverTest extends TestCase
         $clientMessage->setProperties(['key' => 'val']);
         $clientMessage->setContentType('ContentType');
         $clientMessage->setExpire(123);
-        $clientMessage->setPriority(MessagePriority::VERY_HIGH);
-        $clientMessage->setDelay(432);
         $clientMessage->setMessageId('MessageId');
         $clientMessage->setTimestamp(1000);
         $clientMessage->setReplyTo('theReplyTo');
         $clientMessage->setCorrelationId('theCorrelationId');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createMessage')
             ->willReturn(new AmqpMessage())
         ;
 
-        $driver = new RabbitMqDriver(
+        $driver = new AmqpDriver(
             $context,
-            new Config('', '', '', '', '', '', ['delay_plugin_installed' => true]),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -256,11 +183,9 @@ class RabbitMqDriverTest extends TestCase
             'content_type' => 'ContentType',
             'delivery_mode' => 2,
             'expiration' => '123000',
-            'priority' => 4,
         ], $transportMessage->getHeaders());
         $this->assertSame([
             'key' => 'val',
-            'x-delay' => '432000',
         ], $transportMessage->getProperties());
         $this->assertSame('MessageId', $transportMessage->getMessageId());
         $this->assertSame(1000, $transportMessage->getTimestamp());
@@ -268,41 +193,18 @@ class RabbitMqDriverTest extends TestCase
         $this->assertSame('theCorrelationId', $transportMessage->getCorrelationId());
     }
 
-    public function testThrowIfDelayNotSupportedOnConvertClientMessageToTransportMessage()
-    {
-        $clientMessage = new Message();
-        $clientMessage->setDelay(432);
-
-        $context = $this->createPsrContextMock();
-        $context
-            ->expects($this->once())
-            ->method('createMessage')
-            ->willReturn(new AmqpMessage())
-        ;
-
-        $driver = new RabbitMqDriver(
-            $context,
-            new Config('', '', '', '', '', '', ['delay_plugin_installed' => false]),
-            $this->createDummyQueueMetaRegistry()
-        );
-
-        $this->expectException(\LogicException::class);
-        $this->expectExceptionMessage('The message delaying is not supported. In order to use delay feature install RabbitMQ delay plugin.');
-        $driver->createTransportMessage($clientMessage);
-    }
-
     public function testShouldSendMessageToRouter()
     {
         $topic = new AmqpTopic('');
         $transportMessage = new AmqpMessage();
 
-        $producer = $this->createPsrProducerMock();
+        $producer = $this->createAmqpProducerMock();
         $producer
             ->expects($this->once())
             ->method('send')
             ->with($this->identicalTo($topic), $this->identicalTo($transportMessage))
         ;
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createTopic')
@@ -319,9 +221,9 @@ class RabbitMqDriverTest extends TestCase
             ->willReturn($transportMessage)
         ;
 
-        $driver = new RabbitMqDriver(
+        $driver = new AmqpDriver(
             $context,
-            Config::create(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -333,9 +235,9 @@ class RabbitMqDriverTest extends TestCase
 
     public function testShouldThrowExceptionIfTopicParameterIsNotSet()
     {
-        $driver = new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            Config::create(),
+        $driver = new AmqpDriver(
+            $this->createAmqpContextMock(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -350,13 +252,13 @@ class RabbitMqDriverTest extends TestCase
         $queue = new AmqpQueue('');
         $transportMessage = new AmqpMessage();
 
-        $producer = $this->createPsrProducerMock();
+        $producer = $this->createAmqpProducerMock();
         $producer
             ->expects($this->once())
             ->method('send')
             ->with($this->identicalTo($queue), $this->identicalTo($transportMessage))
         ;
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createQueue')
@@ -373,72 +275,24 @@ class RabbitMqDriverTest extends TestCase
             ->willReturn($transportMessage)
         ;
 
-        $driver = new RabbitMqDriver(
+        $driver = new AmqpDriver(
             $context,
-            Config::create(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
         $message = new Message();
         $message->setProperty(Config::PARAMETER_PROCESSOR_NAME, 'processor');
         $message->setProperty(Config::PARAMETER_PROCESSOR_QUEUE_NAME, 'aFooQueue');
-
-        $driver->sendToProcessor($message);
-    }
-
-    public function testShouldSendMessageToDelayExchangeIfDelaySet()
-    {
-        $queue = new AmqpQueue('');
-        $delayTopic = new AmqpTopic('');
-        $transportMessage = new AmqpMessage();
-
-        $producer = $this->createPsrProducerMock();
-        $producer
-            ->expects($this->once())
-            ->method('send')
-            ->with($this->identicalTo($delayTopic), $this->identicalTo($transportMessage))
-        ;
-        $context = $this->createPsrContextMock();
-        $context
-            ->expects($this->once())
-            ->method('createQueue')
-            ->willReturn($queue)
-        ;
-        $context
-            ->expects($this->once())
-            ->method('createTopic')
-            ->willReturn($delayTopic)
-        ;
-        $context
-            ->expects($this->once())
-            ->method('createProducer')
-            ->willReturn($producer)
-        ;
-        $context
-            ->expects($this->once())
-            ->method('createMessage')
-            ->willReturn($transportMessage)
-        ;
-
-        $driver = new RabbitMqDriver(
-            $context,
-            new Config('', '', '', '', '', '', ['delay_plugin_installed' => true]),
-            $this->createDummyQueueMetaRegistry()
-        );
-
-        $message = new Message();
-        $message->setProperty(Config::PARAMETER_PROCESSOR_NAME, 'processor');
-        $message->setProperty(Config::PARAMETER_PROCESSOR_QUEUE_NAME, 'aFooQueue');
-        $message->setDelay(10);
 
         $driver->sendToProcessor($message);
     }
 
     public function testShouldThrowExceptionIfProcessorNameParameterIsNotSet()
     {
-        $driver = new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            Config::create(),
+        $driver = new AmqpDriver(
+            $this->createAmqpContextMock(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -450,9 +304,9 @@ class RabbitMqDriverTest extends TestCase
 
     public function testShouldThrowExceptionIfProcessorQueueNameParameterIsNotSet()
     {
-        $driver = new RabbitMqDriver(
-            $this->createPsrContextMock(),
-            Config::create(),
+        $driver = new AmqpDriver(
+            $this->createAmqpContextMock(),
+            $this->createDummyConfig(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -465,66 +319,14 @@ class RabbitMqDriverTest extends TestCase
         $driver->sendToProcessor($message);
     }
 
-    public function testShouldSetupBrokerWhenDelayPluginNotInstalled()
-    {
-        $routerTopic = new AmqpTopic('');
-        $routerQueue = new AmqpQueue('');
-
-        $processorQueue = new AmqpQueue('');
-        $delayTopic = new AmqpTopic('');
-
-        $context = $this->createPsrContextMock();
-        // setup router
-        $context
-            ->expects($this->at(0))
-            ->method('createTopic')
-            ->willReturn($routerTopic)
-        ;
-        $context
-            ->expects($this->at(1))
-            ->method('createQueue')
-            ->willReturn($routerQueue)
-        ;
-        $context
-            ->expects($this->at(2))
-            ->method('declareTopic')
-            ->with($this->identicalTo($routerTopic))
-        ;
-        $context
-            ->expects($this->at(3))
-            ->method('declareQueue')
-            ->with($this->identicalTo($routerQueue))
-        ;
-        $context
-            ->expects($this->at(4))
-            ->method('bind')
-            ->with($this->isInstanceOf(AmqpBind::class))
-        ;
-        // setup processor queue
-        $context
-            ->expects($this->at(5))
-            ->method('createQueue')
-            ->willReturn($processorQueue)
-        ;
-
-        $config = Config::create('', '', '', '', '', '', ['delay_plugin_installed' => false]);
-
-        $meta = new QueueMetaRegistry($config, ['default' => []]);
-
-        $driver = new RabbitMqDriver($context, $config, $meta);
-
-        $driver->setupBroker();
-    }
-
     public function testShouldSetupBroker()
     {
         $routerTopic = new AmqpTopic('');
         $routerQueue = new AmqpQueue('');
 
         $processorQueue = new AmqpQueue('');
-        $delayTopic = new AmqpTopic('');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         // setup router
         $context
             ->expects($this->at(0))
@@ -562,33 +364,16 @@ class RabbitMqDriverTest extends TestCase
             ->method('declareQueue')
             ->with($this->identicalTo($processorQueue))
         ;
-        $context
-            ->expects($this->at(7))
-            ->method('createQueue')
-            ->willReturn($processorQueue)
-        ;
-        $context
-            ->expects($this->at(8))
-            ->method('createTopic')
-            ->willReturn($delayTopic)
-        ;
-        $context
-            ->expects($this->at(9))
-            ->method('declareTopic')
-            ->with($this->identicalTo($delayTopic))
-        ;
 
-        $context
-            ->expects($this->at(10))
-            ->method('bind')
-            ->with($this->isInstanceOf(AmqpBind::class))
-        ;
+        $meta = new QueueMetaRegistry($this->createDummyConfig(), [
+            'default' => [],
+        ]);
 
-        $config = Config::create('', '', '', '', '', '', ['delay_plugin_installed' => true]);
-
-        $meta = new QueueMetaRegistry($config, ['default' => []]);
-
-        $driver = new RabbitMqDriver($context, $config, $meta);
+        $driver = new AmqpDriver(
+            $context,
+            $this->createDummyConfig(),
+            $meta
+        );
 
         $driver->setupBroker();
     }
@@ -596,17 +381,17 @@ class RabbitMqDriverTest extends TestCase
     /**
      * @return \PHPUnit_Framework_MockObject_MockObject|AmqpContext
      */
-    private function createPsrContextMock()
+    private function createAmqpContextMock()
     {
         return $this->createMock(AmqpContext::class);
     }
 
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|PsrProducer
+     * @return \PHPUnit_Framework_MockObject_MockObject|AmqpProducer
      */
-    private function createPsrProducerMock()
+    private function createAmqpProducerMock()
     {
-        return $this->createMock(PsrProducer::class);
+        return $this->createMock(AmqpProducer::class);
     }
 
     /**
@@ -614,10 +399,19 @@ class RabbitMqDriverTest extends TestCase
      */
     private function createDummyQueueMetaRegistry()
     {
-        $registry = new QueueMetaRegistry(Config::create('aPrefix'), []);
+        $registry = new QueueMetaRegistry($this->createDummyConfig(), []);
+        $registry->add('default');
         $registry->add('aFooQueue');
         $registry->add('aBarQueue', 'aBarQueue');
 
         return $registry;
+    }
+
+    /**
+     * @return Config
+     */
+    private function createDummyConfig()
+    {
+        return Config::create('aPrefix');
     }
 }
