+++ pkg/enqueue/Tests/Client/Amqp/RabbitMqDriverTest.php
@@ -1,44 +1,51 @@
 <?php
 
-namespace Enqueue\AmqpExt\Tests\Client;
+namespace Enqueue\Tests\Client\Amqp;
 
-use Enqueue\AmqpExt\AmqpContext;
-use Enqueue\AmqpExt\Client\AmqpDriver;
+use Enqueue\Client\Amqp\AmqpDriver;
+use Enqueue\Client\Amqp\RabbitMqDriver;
 use Enqueue\Client\Config;
 use Enqueue\Client\DriverInterface;
 use Enqueue\Client\Message;
+use Enqueue\Client\MessagePriority;
 use Enqueue\Client\Meta\QueueMetaRegistry;
 use Enqueue\Test\ClassExtensionTrait;
-use Interop\Amqp\Impl\AmqpBind;
+use Interop\Amqp\AmqpBind;
+use Interop\Amqp\AmqpContext;
+use Interop\Amqp\AmqpProducer;
 use Interop\Amqp\Impl\AmqpMessage;
 use Interop\Amqp\Impl\AmqpQueue;
 use Interop\Amqp\Impl\AmqpTopic;
-use Interop\Queue\PsrProducer;
 use PHPUnit\Framework\TestCase;
 
-class AmqpDriverTest extends TestCase
+class RabbitMqDriverTest extends TestCase
 {
     use ClassExtensionTrait;
 
     public function testShouldImplementsDriverInterface()
     {
-        $this->assertClassImplements(DriverInterface::class, AmqpDriver::class);
+        $this->assertClassImplements(DriverInterface::class, RabbitMqDriver::class);
+    }
+
+    public function testShouldExtendsAmqpDriverClass()
+    {
+        $this->assertClassExtends(AmqpDriver::class, RabbitMqDriver::class);
     }
 
     public function testCouldBeConstructedWithRequiredArguments()
     {
-        new AmqpDriver(
-            $this->createPsrContextMock(),
-            $this->createDummyConfig(),
+        new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            Config::create(),
             $this->createDummyQueueMetaRegistry()
         );
     }
 
     public function testShouldReturnConfigObject()
     {
-        $config = $this->createDummyConfig();
+        $config = Config::create();
 
-        $driver = new AmqpDriver($this->createPsrContextMock(), $config, $this->createDummyQueueMetaRegistry());
+        $driver = new RabbitMqDriver($this->createAmqpContextMock(), $config, $this->createDummyQueueMetaRegistry());
 
         $this->assertSame($config, $driver->getConfig());
     }
@@ -47,7 +54,7 @@ class AmqpDriverTest extends TestCase
     {
         $expectedQueue = new AmqpQueue('aName');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createQueue')
@@ -55,7 +62,7 @@ class AmqpDriverTest extends TestCase
             ->willReturn($expectedQueue)
         ;
 
-        $driver = new AmqpDriver($context, $this->createDummyConfig(), $this->createDummyQueueMetaRegistry());
+        $driver = new AmqpDriver($context, Config::create(), $this->createDummyQueueMetaRegistry());
 
         $queue = $driver->createQueue('aFooQueue');
 
@@ -69,7 +76,7 @@ class AmqpDriverTest extends TestCase
     {
         $expectedQueue = new AmqpQueue('aName');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createQueue')
@@ -77,7 +84,7 @@ class AmqpDriverTest extends TestCase
             ->willReturn($expectedQueue)
         ;
 
-        $driver = new AmqpDriver($context, $this->createDummyConfig(), $this->createDummyQueueMetaRegistry());
+        $driver = new AmqpDriver($context, Config::create(), $this->createDummyQueueMetaRegistry());
 
         $queue = $driver->createQueue('aBarQueue');
 
@@ -90,16 +97,18 @@ class AmqpDriverTest extends TestCase
         $transportMessage->setBody('body');
         $transportMessage->setHeaders(['hkey' => 'hval']);
         $transportMessage->setProperties(['key' => 'val']);
+        $transportMessage->setProperty('x-delay', '5678000');
         $transportMessage->setHeader('content_type', 'ContentType');
         $transportMessage->setHeader('expiration', '12345000');
+        $transportMessage->setHeader('priority', 3);
         $transportMessage->setMessageId('MessageId');
         $transportMessage->setTimestamp(1000);
         $transportMessage->setReplyTo('theReplyTo');
         $transportMessage->setCorrelationId('theCorrelationId');
 
-        $driver = new AmqpDriver(
-            $this->createPsrContextMock(),
-            $this->createDummyConfig(),
+        $driver = new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            new Config('', '', '', '', '', '', ['delay_plugin_installed' => true]),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -111,6 +120,7 @@ class AmqpDriverTest extends TestCase
             'hkey' => 'hval',
             'content_type' => 'ContentType',
             'expiration' => '12345000',
+            'priority' => 3,
             'message_id' => 'MessageId',
             'timestamp' => 1000,
             'reply_to' => 'theReplyTo',
@@ -118,13 +128,33 @@ class AmqpDriverTest extends TestCase
         ], $clientMessage->getHeaders());
         $this->assertSame([
             'key' => 'val',
+            'x-delay' => '5678000',
         ], $clientMessage->getProperties());
         $this->assertSame('MessageId', $clientMessage->getMessageId());
         $this->assertSame(12345, $clientMessage->getExpire());
+        $this->assertSame(5678, $clientMessage->getDelay());
         $this->assertSame('ContentType', $clientMessage->getContentType());
         $this->assertSame(1000, $clientMessage->getTimestamp());
+        $this->assertSame(MessagePriority::HIGH, $clientMessage->getPriority());
         $this->assertSame('theReplyTo', $clientMessage->getReplyTo());
         $this->assertSame('theCorrelationId', $clientMessage->getCorrelationId());
+    }
+
+    public function testShouldThrowExceptionIfXDelayIsNotNumeric()
+    {
+        $transportMessage = new AmqpMessage();
+        $transportMessage->setProperty('x-delay', 'is-not-numeric');
+
+        $driver = new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            Config::create(),
+            $this->createDummyQueueMetaRegistry()
+        );
+
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage('x-delay header is not numeric. "is-not-numeric"');
+
+        $driver->createClientMessage($transportMessage);
     }
 
     public function testShouldThrowExceptionIfExpirationIsNotNumeric()
@@ -132,9 +162,9 @@ class AmqpDriverTest extends TestCase
         $transportMessage = new AmqpMessage();
         $transportMessage->setHeader('expiration', 'is-not-numeric');
 
-        $driver = new AmqpDriver(
-            $this->createPsrContextMock(),
-            $this->createDummyConfig(),
+        $driver = new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            Config::create(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -142,6 +172,47 @@ class AmqpDriverTest extends TestCase
         $this->expectExceptionMessage('expiration header is not numeric. "is-not-numeric"');
 
         $driver->createClientMessage($transportMessage);
+    }
+
+    public function testShouldThrowExceptionIfCantConvertTransportPriorityToClientPriority()
+    {
+        $transportMessage = new AmqpMessage();
+        $transportMessage->setHeader('priority', 'unknown');
+
+        $driver = new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            Config::create(),
+            $this->createDummyQueueMetaRegistry()
+        );
+
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage('Cant convert transport priority to client: "unknown"');
+
+        $driver->createClientMessage($transportMessage);
+    }
+
+    public function testShouldThrowExceptionIfCantConvertClientPriorityToTransportPriority()
+    {
+        $clientMessage = new Message();
+        $clientMessage->setPriority('unknown');
+
+        $context = $this->createAmqpContextMock();
+        $context
+            ->expects($this->once())
+            ->method('createMessage')
+            ->willReturn(new AmqpMessage())
+        ;
+
+        $driver = new RabbitMqDriver(
+            $context,
+            Config::create(),
+            $this->createDummyQueueMetaRegistry()
+        );
+
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage('Given priority could not be converted to client\'s one. Got: unknown');
+
+        $driver->createTransportMessage($clientMessage);
     }
 
     public function testShouldConvertClientMessageToTransportMessage()
@@ -152,21 +223,23 @@ class AmqpDriverTest extends TestCase
         $clientMessage->setProperties(['key' => 'val']);
         $clientMessage->setContentType('ContentType');
         $clientMessage->setExpire(123);
+        $clientMessage->setPriority(MessagePriority::VERY_HIGH);
+        $clientMessage->setDelay(432);
         $clientMessage->setMessageId('MessageId');
         $clientMessage->setTimestamp(1000);
         $clientMessage->setReplyTo('theReplyTo');
         $clientMessage->setCorrelationId('theCorrelationId');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createMessage')
             ->willReturn(new AmqpMessage())
         ;
 
-        $driver = new AmqpDriver(
+        $driver = new RabbitMqDriver(
             $context,
-            $this->createDummyConfig(),
+            new Config('', '', '', '', '', '', ['delay_plugin_installed' => true]),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -183,9 +256,11 @@ class AmqpDriverTest extends TestCase
             'content_type' => 'ContentType',
             'delivery_mode' => 2,
             'expiration' => '123000',
+            'priority' => 4,
         ], $transportMessage->getHeaders());
         $this->assertSame([
             'key' => 'val',
+            'x-delay' => '432000',
         ], $transportMessage->getProperties());
         $this->assertSame('MessageId', $transportMessage->getMessageId());
         $this->assertSame(1000, $transportMessage->getTimestamp());
@@ -193,18 +268,41 @@ class AmqpDriverTest extends TestCase
         $this->assertSame('theCorrelationId', $transportMessage->getCorrelationId());
     }
 
+    public function testThrowIfDelayNotSupportedOnConvertClientMessageToTransportMessage()
+    {
+        $clientMessage = new Message();
+        $clientMessage->setDelay(432);
+
+        $context = $this->createAmqpContextMock();
+        $context
+            ->expects($this->once())
+            ->method('createMessage')
+            ->willReturn(new AmqpMessage())
+        ;
+
+        $driver = new RabbitMqDriver(
+            $context,
+            new Config('', '', '', '', '', '', ['delay_plugin_installed' => false]),
+            $this->createDummyQueueMetaRegistry()
+        );
+
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage('The message delaying is not supported. In order to use delay feature install RabbitMQ delay plugin.');
+        $driver->createTransportMessage($clientMessage);
+    }
+
     public function testShouldSendMessageToRouter()
     {
         $topic = new AmqpTopic('');
         $transportMessage = new AmqpMessage();
 
-        $producer = $this->createPsrProducerMock();
+        $producer = $this->createAmqpProducerMock();
         $producer
             ->expects($this->once())
             ->method('send')
             ->with($this->identicalTo($topic), $this->identicalTo($transportMessage))
         ;
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createTopic')
@@ -221,9 +319,9 @@ class AmqpDriverTest extends TestCase
             ->willReturn($transportMessage)
         ;
 
-        $driver = new AmqpDriver(
+        $driver = new RabbitMqDriver(
             $context,
-            $this->createDummyConfig(),
+            Config::create(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -235,9 +333,9 @@ class AmqpDriverTest extends TestCase
 
     public function testShouldThrowExceptionIfTopicParameterIsNotSet()
     {
-        $driver = new AmqpDriver(
-            $this->createPsrContextMock(),
-            $this->createDummyConfig(),
+        $driver = new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            Config::create(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -252,13 +350,13 @@ class AmqpDriverTest extends TestCase
         $queue = new AmqpQueue('');
         $transportMessage = new AmqpMessage();
 
-        $producer = $this->createPsrProducerMock();
+        $producer = $this->createAmqpProducerMock();
         $producer
             ->expects($this->once())
             ->method('send')
             ->with($this->identicalTo($queue), $this->identicalTo($transportMessage))
         ;
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
         $context
             ->expects($this->once())
             ->method('createQueue')
@@ -275,9 +373,9 @@ class AmqpDriverTest extends TestCase
             ->willReturn($transportMessage)
         ;
 
-        $driver = new AmqpDriver(
+        $driver = new RabbitMqDriver(
             $context,
-            $this->createDummyConfig(),
+            Config::create(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -288,11 +386,59 @@ class AmqpDriverTest extends TestCase
         $driver->sendToProcessor($message);
     }
 
+    public function testShouldSendMessageToDelayExchangeIfDelaySet()
+    {
+        $queue = new AmqpQueue('');
+        $delayTopic = new AmqpTopic('');
+        $transportMessage = new AmqpMessage();
+
+        $producer = $this->createAmqpProducerMock();
+        $producer
+            ->expects($this->once())
+            ->method('send')
+            ->with($this->identicalTo($delayTopic), $this->identicalTo($transportMessage))
+        ;
+        $context = $this->createAmqpContextMock();
+        $context
+            ->expects($this->once())
+            ->method('createQueue')
+            ->willReturn($queue)
+        ;
+        $context
+            ->expects($this->once())
+            ->method('createTopic')
+            ->willReturn($delayTopic)
+        ;
+        $context
+            ->expects($this->once())
+            ->method('createProducer')
+            ->willReturn($producer)
+        ;
+        $context
+            ->expects($this->once())
+            ->method('createMessage')
+            ->willReturn($transportMessage)
+        ;
+
+        $driver = new RabbitMqDriver(
+            $context,
+            new Config('', '', '', '', '', '', ['delay_plugin_installed' => true]),
+            $this->createDummyQueueMetaRegistry()
+        );
+
+        $message = new Message();
+        $message->setProperty(Config::PARAMETER_PROCESSOR_NAME, 'processor');
+        $message->setProperty(Config::PARAMETER_PROCESSOR_QUEUE_NAME, 'aFooQueue');
+        $message->setDelay(10);
+
+        $driver->sendToProcessor($message);
+    }
+
     public function testShouldThrowExceptionIfProcessorNameParameterIsNotSet()
     {
-        $driver = new AmqpDriver(
-            $this->createPsrContextMock(),
-            $this->createDummyConfig(),
+        $driver = new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            Config::create(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -304,9 +450,9 @@ class AmqpDriverTest extends TestCase
 
     public function testShouldThrowExceptionIfProcessorQueueNameParameterIsNotSet()
     {
-        $driver = new AmqpDriver(
-            $this->createPsrContextMock(),
-            $this->createDummyConfig(),
+        $driver = new RabbitMqDriver(
+            $this->createAmqpContextMock(),
+            Config::create(),
             $this->createDummyQueueMetaRegistry()
         );
 
@@ -319,14 +465,65 @@ class AmqpDriverTest extends TestCase
         $driver->sendToProcessor($message);
     }
 
-    public function testShouldSetupBroker()
+    public function testShouldSetupBrokerWhenDelayPluginNotInstalled()
     {
         $routerTopic = new AmqpTopic('');
         $routerQueue = new AmqpQueue('');
 
         $processorQueue = new AmqpQueue('');
 
-        $context = $this->createPsrContextMock();
+        $context = $this->createAmqpContextMock();
+        // setup router
+        $context
+            ->expects($this->at(0))
+            ->method('createTopic')
+            ->willReturn($routerTopic)
+        ;
+        $context
+            ->expects($this->at(1))
+            ->method('createQueue')
+            ->willReturn($routerQueue)
+        ;
+        $context
+            ->expects($this->at(2))
+            ->method('declareTopic')
+            ->with($this->identicalTo($routerTopic))
+        ;
+        $context
+            ->expects($this->at(3))
+            ->method('declareQueue')
+            ->with($this->identicalTo($routerQueue))
+        ;
+        $context
+            ->expects($this->at(4))
+            ->method('bind')
+            ->with($this->isInstanceOf(AmqpBind::class))
+        ;
+        // setup processor queue
+        $context
+            ->expects($this->at(5))
+            ->method('createQueue')
+            ->willReturn($processorQueue)
+        ;
+
+        $config = Config::create('', '', '', '', '', '', ['delay_plugin_installed' => false]);
+
+        $meta = new QueueMetaRegistry($config, ['default' => []]);
+
+        $driver = new RabbitMqDriver($context, $config, $meta);
+
+        $driver->setupBroker();
+    }
+
+    public function testShouldSetupBroker()
+    {
+        $routerTopic = new AmqpTopic('');
+        $routerQueue = new AmqpQueue('');
+
+        $processorQueue = new AmqpQueue('');
+        $delayTopic = new AmqpTopic('');
+
+        $context = $this->createAmqpContextMock();
         // setup router
         $context
             ->expects($this->at(0))
@@ -364,16 +561,33 @@ class AmqpDriverTest extends TestCase
             ->method('declareQueue')
             ->with($this->identicalTo($processorQueue))
         ;
+        $context
+            ->expects($this->at(7))
+            ->method('createQueue')
+            ->willReturn($processorQueue)
+        ;
+        $context
+            ->expects($this->at(8))
+            ->method('createTopic')
+            ->willReturn($delayTopic)
+        ;
+        $context
+            ->expects($this->at(9))
+            ->method('declareTopic')
+            ->with($this->identicalTo($delayTopic))
+        ;
 
-        $meta = new QueueMetaRegistry($this->createDummyConfig(), [
-            'default' => [],
-        ]);
+        $context
+            ->expects($this->at(10))
+            ->method('bind')
+            ->with($this->isInstanceOf(AmqpBind::class))
+        ;
 
-        $driver = new AmqpDriver(
-            $context,
-            $this->createDummyConfig(),
-            $meta
-        );
+        $config = Config::create('', '', '', '', '', '', ['delay_plugin_installed' => true]);
+
+        $meta = new QueueMetaRegistry($config, ['default' => []]);
+
+        $driver = new RabbitMqDriver($context, $config, $meta);
 
         $driver->setupBroker();
     }
@@ -381,17 +595,17 @@ class AmqpDriverTest extends TestCase
     /**
      * @return \PHPUnit_Framework_MockObject_MockObject|AmqpContext
      */
-    private function createPsrContextMock()
+    private function createAmqpContextMock()
     {
         return $this->createMock(AmqpContext::class);
     }
 
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|PsrProducer
+     * @return \PHPUnit_Framework_MockObject_MockObject|AmqpProducer
      */
-    private function createPsrProducerMock()
+    private function createAmqpProducerMock()
     {
-        return $this->createMock(PsrProducer::class);
+        return $this->createMock(AmqpProducer::class);
     }
 
     /**
@@ -399,19 +613,10 @@ class AmqpDriverTest extends TestCase
      */
     private function createDummyQueueMetaRegistry()
     {
-        $registry = new QueueMetaRegistry($this->createDummyConfig(), []);
-        $registry->add('default');
+        $registry = new QueueMetaRegistry(Config::create('aPrefix'), []);
         $registry->add('aFooQueue');
         $registry->add('aBarQueue', 'aBarQueue');
 
         return $registry;
-    }
-
-    /**
-     * @return Config
-     */
-    private function createDummyConfig()
-    {
-        return Config::create('aPrefix');
     }
 }
